requirements-dev.txt
# Dev dependencies for TRMNL template screenshot tests
python-liquid>=1.12.0
playwright>=1.40.0
pytest>=8.0.0
pytest-playwright>=0.4.0
pytest-asyncio>=0.25.0
httpx>=0.28.0
